found a way also to make it with player
basically, instead of adding collision in ai, there is a way also to make in the player itself

also made sure to set players flashlight's intensity to a number, so when flickering is stopping, the light doesnt accidently stop at 0, which would happen, which is fixed.

ai detecting the player is still optinal. so there is 2 ways for the dectection flickering