interaction on indiviual computer is working (kind of)
added 2 material for correct and incorrect displayer

interaction is working, but the only thing is that it can change the screen to incorrect screen when the loading is  not fully loaded.

also when interacting with the computer, it will not show "correct" screen until you got the code right. so if the correct code is 123 and you interact with 123. 1 will show incorrect, 2 will show incorrect, 3 will show correct.